Fix large logical max (greater than i32) improperly handled for field_value() (#9)
## Description

HID spec indicates that LogicalMaximum should be treated as signed (i32)
when LogicalMinimum is negative, but unsigned (u32) when LogicalMinum is
not negative. Prior to this modification, the field_value() routine was
always treating LogicalMaximum as i32, resulting in incorrect handling
of some fields. This change addresses that, and adds a test to catch it.

- [x] Impacts functionality?
  - Fixes a bug in field_value() computation
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested
Verified in QEMU, which has a USB keyboard implementation that reports
LogicalMinimum(0), LogicalMaximum(0xffffffff) that triggers this issue.

This PR also incorporates unit tests to reproduce the issue, and these
tests pass.

## Integration Instructions
N/A