Refine report parsing to allow a non-spec compliant report to be handled without aborting processing of the whole descriptor. (#25)
Refine report parsing to allow a non-spec compliant report to be handled
without aborting processing of the whole descriptor.

## Description

This allows handling descriptors that are not completely per HID spec;
this allows partial support for devices where some reports are not
valid, but some are. This gives more flexibility in supporting hardware
with various degrees of adherence to spec requirements.

- [x] Impacts functionality?
- If an error is detected in a report descriptor, only the affected
reports are not parsed. Other reports that are valid within the same
descriptor are still parsed and returned. Parsing returns a list of any
errors determined along with the ids (if any) of the affected reports.
- [ ] Impacts security?
- [x] Breaking change?
- The main parsing API (`ReportDescriptor::parse(&[u8]) ->
Result<ReportDescriptor, ReportDescriptorError> does not change, but the
definition of ReportDescriptor expands to include "bad descriptor" lists
to report parsing failures, and error is no longer returned for the
whole descriptor for all errors. Invalid item tokenization will still
cause the whole descriptor parse to fail, but if an error is localized
(like missing report count), then the remaining unaffected reports are
returned.
- [x] Includes tests?
  - Tests and examples updated to accommodate the change.
- [ ] Includes documentation?

## How This Was Tested

Unit tests updated and passed, example parser behavior verified on good
and bad descriptors.

## Integration Instructions
N/A unless the application was previously relying on descriptor parse
failure semantics in some way, in which case error cases may need to be
revisited.